Add weather list page

+++ learnflutter/lib/exercise/weather_app/models/weather_result.dart
@@ -134,7 +134,7 @@ class WeatherMainObject {
 }
 
 class WindObject {
-  int? speed;
+  double? speed;
   int? deg;
   double? gust;
